# recipes/service/spec/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "recipes API",
    "version": "1.0.0",
    "description": "API specification for the recipes service"
  },
  "paths": {
...
    "/recipes/{id}/versions": {
      "get": {
        "summary": "List version history for a recipe (ordered by createdAt).",
        "operationId": "listRecipeVersions",
        "tags": [
          "no-auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "responses": {
          "200": {
            "description": "Version history for the recipe (ordered by createdAt).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/recipe-version"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - recipe is private and caller is not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new recipe version (recorded history entry). Admin only.",
        "operationId": "createRecipeVersion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Content for the new version.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/content"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New RecipeVersion (set as latest).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe-version"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/versions/latest": {
      "put": {
        "summary": "Update the latest version in place (no new version created). Admin only.",
        "operationId": "updateRecipeVersionLatest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Version content (full or partial); same shape as RecipeVersion content.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/content"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated RecipeVersion (the latest).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe-version"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A short, machine-readable error code, for when HTTP status codes are not sufficient."
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the error.",
            "example": "The requested resource could not be found."
          }
        }
      },
      "content": {
        "type": "object",
        "description": "Version body — structured ingredients and markdown instructions",
        "properties": {
          "ingredients": {
            "type": "array",
            "description": "List of ingredients; quantity and unit are strings (e.g. \"2+1/3\", \"cups\")",
            "items": {
              "type": "object",
              "required": [
                "name",
                "quantity",
                "unit"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Ingredient name",
                  "example": "All-purpose flour"
                },
                "quantity": {
                  "type": "string",
                  "description": "Quantity as string to support values like \"1/3\", \"2+1/3\"",
                  "example": "2+1/3"
                },
                "unit": {
                  "type": "string",
                  "description": "Unit of measure (unrestricted string)",
                  "example": "cups"
                }
              }
            },
            "example": [
              {
                "name": "All-purpose flour",
                "quantity": "2",
                "unit": "cups"
              },
              {
                "name": "Butter",
                "quantity": "1/2",
                "unit": "cup"
              }
            ]
          },
          "instructionsMarkdown": {
            "type": "string",
            "description": "Recipe instructions in markdown",
            "example": "1. Preheat oven to 350°F.\n2. Mix dry ingredients..."
          }
        },
        "required": [
          "ingredients",
          "instructionsMarkdown"
        ],
        "example": {
          "ingredients": [
            {
              "name": "All-purpose flour",
              "quantity": "2",
              "unit": "cups"
            }
          ],
          "instructionsMarkdown": "1. Preheat oven to 350°F.\n2. Mix dry ingredients."
        }
      },
      "recipe-version": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for this recipe version",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeId": {
            "type": "string",
            "description": "Id of the recipe this version belongs to",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "content": {
            "type": "object",
            "description": "Version body — structured ingredients and markdown instructions",
            "properties": {
              "ingredients": {
                "type": "array",
                "description": "List of ingredients; quantity and unit are strings (e.g. \"2+1/3\", \"cups\")",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "quantity",
                    "unit"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Ingredient name",
                      "example": "All-purpose flour"
                    },
                    "quantity": {
                      "type": "string",
                      "description": "Quantity as string to support values like \"1/3\", \"2+1/3\"",
                      "example": "2+1/3"
                    },
                    "unit": {
                      "type": "string",
                      "description": "Unit of measure (unrestricted string)",
                      "example": "cups"
                    }
                  }
                },
                "example": [
                  {
                    "name": "All-purpose flour",
                    "quantity": "2",
                    "unit": "cups"
                  },
                  {
                    "name": "Butter",
                    "quantity": "1/2",
                    "unit": "cup"
                  }
                ]
              },
              "instructionsMarkdown": {
                "type": "string",
                "description": "Recipe instructions in markdown",
                "example": "1. Preheat oven to 350°F.\n2. Mix dry ingredients..."
              }
            },
            "required": [
              "ingredients",
              "instructionsMarkdown"
            ],
            "example": {
              "ingredients": [
                {
                  "name": "All-purpose flour",
                  "quantity": "2",
                  "unit": "cups"
                }
              ],
              "instructionsMarkdown": "1. Preheat oven to 350°F.\n2. Mix dry ingredients."
            }
          },
          "isLatest": {
            "type": "boolean",
            "description": "Whether this version is the current one for the recipe (only one per recipe is true)",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator of this version",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this version was created",
            "example": "2023-01-15T14:30:00Z"
          }
        },
        "required": [
          "id",
          "recipeId",
          "content",
          "isLatest",
          "createdBy",
          "createdAt"
        ]
      },
...
    }
  }
}